+++ nequip/data/_key_registry.py
@@ -20,6 +20,7 @@ _DEFAULT_LONG_FIELDS: Set[str] = {
     _keys.DATASET_KEY,
     _keys.TOTAL_CHARGE_KEY,
     _keys.TOTAL_SPIN_KEY,
+    _keys.HEAD_KEY,
 }
 _DEFAULT_GRAPH_FIELDS: Set[str] = {
     _keys.TOTAL_ENERGY_KEY,
@@ -35,6 +36,8 @@ _DEFAULT_GRAPH_FIELDS: Set[str] = {
     _keys.TOTAL_MAGMOM_KEY,
     _keys.POLARIZATION_KEY,
     _keys.DIELECTRIC_KEY,
+    _keys.FRAME_SUBSET_KEY,
+    _keys.HEAD_KEY,
 }
 _DEFAULT_NODE_FIELDS: Set[str] = {
     _keys.POSITIONS_KEY,