+++ nequip/data/transforms/__init__.py
@@ -14,6 +14,7 @@ from .stress_utils import (
 from .cell_utils import (
     NonPeriodicCellTransform,
 )
+from .head_stamper import HeadStamper
 
 __all__ = [
     "ChemicalSpeciesToAtomTypeMapper",
@@ -25,4 +26,5 @@ __all__ = [
     "StressSignFlipTransform",
     "AddNaNStressTransform",
     "NonPeriodicCellTransform",
+    "HeadStamper",
 ]