+++ nequip/nn/__init__.py
@@ -16,6 +16,7 @@ from .misc import Concat, ApplyFactor, SaveForOutput
 from .utils import scatter, tp_path_exists, with_edge_vectors_, with_edge_type_
 from .model_modifier_utils import model_modifier, replace_submodules
 from .norm import AvgNumNeighborsNorm
+from .multihead import MultiHeadReadout
 
 __all__ = [
     "GraphModel",
@@ -42,4 +43,5 @@ __all__ = [
     "model_modifier",
     "replace_submodules",
     "AvgNumNeighborsNorm",
+    "MultiHeadReadout",
 ]